pyproject.toml
[tool.poetry]
name = "krunner-kdbx"
version = "0.1.0"
description = "KRunner plugin for KDBX database (KeePass/KeePassXC)"
authors = ["vonabarak <github@vonabarak.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dbus-python = "^1.3.2"
setproctitle = "^1.3.3"
pyotp = "^2.9.0"
pykeepass = "^4.0.7"

[tool.poetry.scripts]
krunner-kdbx = "krunner_kdbx.__main__:main"
krunner-kdbx-helper = "krunner_kdbx.helper:open_db"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

